refactor mapstructure
