+++ src/main/java/demo/codeexample/logger/domain/LoggerEntity.java
@@ -1,4 +1,4 @@
-package demo.codeexample.logger;
+package demo.codeexample.logger.domain;
 
 import demo.codeexample.enums.LoggerAction;
